streamline promise rejection filter
