+++ packages/utils/src/process/index.ts
@@ -1,3 +1,4 @@
 export * from './debug.js'
+export * from './promiseRejectionFilter.js'
 export * from './resourceLogger.js'
 export * from './retimer.js'
